shoe_wear_exhibition/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
Pillow
numpy
safetensors
accelerate
python-multipart
numpy
rembg